# Specialize broadcast to avoid integer divisions.